move ops to new file
