+++ src/core/core.cxx
@@ -7,3 +7,4 @@ export import :utility;
 export import :frame;
 export import :constants;
 export import :tuple;
+export import :ops;
